+++ Assets/Scripts/FluidDemo/ParticleVisualization.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 
-namespace FluidSimulation
+namespace FluidDemo
 {
     public class ParticleVisualization : MonoBehaviour
     {
