--- Assets/Scripts/FluidDemo/TestWallCreator.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-public class TestWallCreator : MonoBehaviour
-{
-    public float spacing = 1f;
-    public int layersX = 5;
-    public int layersY = 5;
-}
